requirements.txt
# Django Web Framework
Django==5.2.7
asgiref==3.10.0
sqlparse==0.5.3

# WSGI Server for Production
gunicorn==23.0.0

# Static Files Management
whitenoise==6.11.0

# PyTorch for ML Model
torch==2.9.0
torchvision==0.24.0
triton==3.5.0

# Image Processing
Pillow==12.0.0

# Numerical Computing
numpy==2.3.4

# Template Engine (already included with Django but listed explicitly)
Jinja2==3.1.6
MarkupSafe==3.0.3

# Utilities
filelock==3.20.0
fsspec==2025.10.0
packaging==25.0
typing-extensions==4.15.0
setuptools==80.9.0
sympy==1.14.0
mpmath==1.3.0
networkx==3.5

# NVIDIA CUDA Dependencies (for GPU support)
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
